-add a generic class for color some bin files.
-after check emd binaries on all emd and nsk files, by color/tags and find holes or rewrite byte.
     -new version of shaders (better but still not perfect, with many of them corrected manually), but the result it's really better for stage and skill. And also, I extended to the last update of the game where a lot of shaders have been added (a lot for photo's mode).
    -discover last unknow values in TextureUnitState, for filtering and adressMode. (todo do the same check on emo).
    -try a detection and rewrite of uin32 instead uint16 for face index. game seam to accepted.

-one part of checking things for emd, I have redo a analyzeXml of all emm :
   -so, it's was for upgrade shaders entries :
        -making Billboard (like moon or text rotate on a axis)
        -mipmap bias for texture.
        -AlphaTest witch is a alpha rejection, but my fault was to do on pass, instead let do shaders if needed.
        -some shader never change a look because light and camera are fixed for light calculations.

   -I have check all type of emm parameter, so now, it's reste only float and uint, and we have a hidden unknowValue.

-I change nskpack for making bytes split and remerge of Emd and esk (except for the offset adress for target emd). that could help to be sure to have same emd.

-in xenoviewer, add a raySceneQuery for get objects names and material under the mouse position.

-a hyp make me build dyt texture for player color, but the hyp was wrong, but in future we could have player color to be usefull. so I keep it.

-solve bug witch make crash AnalyzerXml sometime.
-solve crash when emb have more than 1000 file like new version of shaders.

-try to force working dir to be relative to the exe, for possiblity to open xenoviwer when double click or drag and drop a emd , nsk or emo files.

+++ LibXenoverse/LibXenoverse/EMDExportFBX.cpp
@@ -242,7 +242,7 @@ void EMDSubmesh::exportFBX(FbxScene *scene, std::vector<ESK::FbxBonesInstance_DB
 	size_t nbTriangles = triangles.size();
 	for (size_t i = 0; i < nbTriangles; i++)
 	{
-		std::vector<unsigned short> &faces = triangles.at(i).faces;
+		std::vector<unsigned int> &faces = triangles.at(i).faces;
 
 
 		// Skinning prepartion part : make a list of bones for triangle.
